// src/api/types/projectId_3631.ts
export interface OrderListResponse {
  /**
   * 流水总额
   */
  allOrderMoney?: number
  /**
   * 绩效营收
   */
  performanceMoney?: number
  /**
   * 工单实际营收
   */
  orderActualIncomeSum?: number
  /**
   * 签约订单列表
   */
  signOrderList?: {
    /**
     * 工单号
     */
    orderNo?: string
    /**
     * 品牌id
     */
    brandId?: number
    /**
     * 品牌名称
     */
    newsContent?: string
    /**
     * 商单类型 1 客户自行下单  2 达人私单  3 不走平台的私单
     */
    quotationType?: number
    /**
     * 工单类型类型：1=视频工单、2=特殊工单
     */
    orderType?: number
    /**
     * 工单状态：1-确认下单信息，2-分配执行人，3-录入执行计划，4-确认大纲，5-确认脚本，6-确认视频初稿，7-发布视频，8-确认收付款，9-财务核账
     */
    orderStatus?: number
    /**
     * 创建人id
     */
    createUserId?: number
    /**
     * 创建人
     */
    createUserName?: number
    /**
     * 商务团队id
     */
    belongGroupId?: number
    /**
     * 商务团队名
     */
    belongGroupName?: string
    /**
     * 订单归属
     */
    belongTeam?: string
    /**
     * 签约经纪人id
     */
    signUserId?: number
    /**
     * 签约经纪人
     */
    signUserName?: string
    /**
     * 签约小组id
     */
    signGroupId?: number
    /**
     * 签约小组
     */
    signGroupName?: string
    /**
     * 签约团队id
     */
    signTeamId?: number
    /**
     * 签约团队
     */
    signTeamName?: string
    /**
     * 账号名
     */
    accountName?: string
    /**
     * 账号id
     */
    accountId?: number
    /**
     * 后台登录链接
     */
    accountLoginUrl?: string
    /**
     * 星图链接
     */
    xingtuIndexUrl?: string
    /**
     * 标签
     */
    sysAccountTag?: string
    /**
     * 平台id
     */
    platId?: number
    /**
     * 平台名
     */
    platName?: string
    /**
     * 档期
     */
    slotDate?: string
    /**
     * 订单流水
     */
    orderMoney?: number
    /**
     * 绩效营收
     */
    performanceMoney?: number
    /**
     * 内容协助  0-否 1-是
     */
    contentAssist?: number
    /**
     * 视频id
     */
    flowId?: number
    /**
     * 视频链接
     */
    url?: string
    /**
     * 视频标题
     */
    titleName?: string
    /**
     * 视频封面链接
     */
    videoCoverUrl?: string
    /**
     * 收付款状态 1-向达人应付 2-向达人应收
     */
    paymentType?: number
    /**
     * 收款状态 1-已付款/已收款 2-未付款/未收款
     */
    paymentResult?: number
    /**
     * 绩效月
     */
    monthMoney?: string
    /**
     * 销售收入
     */
    salesIncomeSum?: number
    /**
     * 工单实际营收
     */
     orderActualIncome?: number
  }[]
  /**
   * 总条数
   */
  total: string
}